requirements.txt
# Core FastAPI and web framework
fastapi==0.115.4
uvicorn[standard]==0.32.1
python-multipart==0.0.12
starlette==0.41.3

# HTTP client and async support
httpx==0.28.1
aiofiles==24.1.0
anyio==4.9.0

# AI and embeddings
openai==1.96.1

# Vector database

# Database (PostgreSQL)
psycopg[binary]

# Document processing
python-docx==1.1.2
python-email==0.0.1

# NLP and text processing (optional, enhanced features)
spacy
# Run: python -m spacy download en_core_web_sm

# Data processing and utilities
pydantic
pydantic-core
python-dotenv==1.0.1
pandas
numpy

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging and monitoring
structlog==25.1.0

# Development and testing (optional)
pytest==8.3.4
pytest-asyncio==0.25.0
black>=24.1.0,<26
isort==5.13.2
langchain-community>=0.0.34

# Backward compatibility - keeping existing working packages
accelerate
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
huggingface-hub==0.33.4
idna==3.6
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
packaging==23.2
pathlib==1.0.1
psutil==7.0.0
python-dateutil==2.8.2
PyYAML==6.0.2
safetensors==0.5.3
six==1.16.0
tqdm==4.66.4
pydot==4.0.1
PyMuPDF==1.26.3
PyMySQL==1.1.1
pyparsing==3.2.3
pytz==2025.2
pyxnat==1.6.3
rdflib==7.1.4
regex==2023.12.25
requests>=2.32.2
scipy
setuptools==80.9.0
simplejson==3.20.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.21.2
torch
traits==7.0.2
transformers
trl==0.19.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
pinecone
scikit-learn>=1.5.0
unstructured
